% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineAtlas-package.R
\docType{package}
\name{ineAtlas-package}
\alias{ineAtlas}
\alias{ineAtlas-package}
\title{ineAtlas: Access to Spanish Household Income Distribution Atlas Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to granular socioeconomic indicators from the Spanish Statistical Office (INE) Household Income Distribution Atlas. The package downloads and processes data from a companion 'GitHub' repository (\url{https://github.com/pablogguz/ineAtlas.data/}) which contains processed versions of the official INE Atlas data. Functions are provided to fetch data at multiple geographic levels (municipalities, districts, and census tracts), including income indicators, demographic characteristics, and inequality metrics. The data repository is updated every year when new releases are published by INE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pablogguz/ineAtlas}
  \item \url{https://pablogguz.github.io/ineAtlas/}
  \item Report bugs at \url{https://github.com/pablogguz/ineAtlas/issues}
}

}
\author{
\strong{Maintainer}: Pablo García Guzmán \email{garciagp@ebrd.com} [copyright holder]

}
\keyword{internal}
